+++ gems/aws-sdk-kendra/lib/aws-sdk-kendra/endpoints.rb
@@ -740,6 +740,20 @@ module Aws::Kendra
       end
     end
 
+    class Retrieve
+      def self.build(context)
+        unless context.config.regional_endpoint
+          endpoint = context.config.endpoint.to_s
+        end
+        Aws::Kendra::EndpointParameters.new(
+          region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
+          use_fips: context.config.use_fips_endpoint,
+          endpoint: endpoint,
+        )
+      end
+    end
+
     class StartDataSourceSyncJob
       def self.build(context)
         unless context.config.regional_endpoint