+++ gems/aws-sdk-chimesdkmessaging/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.21.0 (2023-06-22)
+------------------
+
+* Feature - ChannelMessages can be made visible to sender and intended recipient rather than all channel members with the target attribute. For example, a user can send messages to a bot and receive messages back in a group channel without other members seeing them.
+
 1.20.0 (2023-06-15)
 ------------------